* master: (78 commits)
  [ATLAS][ADD] add support for loading model from memory for atlas
  [ATLAS][CHG] clang-format atlas related code
  [METAL] update KernelName, add inputs and outputs as arguments (#262)
  [ARM] open reverse channel in blob converter
  [ATLAS][BUG] fix memcpy size error when input format is YUV
  [ATLAS][ADD] add dvpp batch support
  [UTILS][FORMAT]统一函数命名规范
  [ATLAS][ADD] add pad value support in ResizeAndPaste
  [ARM] fix inplace reverse channel
  [ARM][BUG] fix half-type model infer error (#274)
  [BENCHMARK][UPD] 1. delete yolov3-tiny.tnnprot; 2. udpate inception_v3.tnnproto;
  [TOOLS][BUG] fix caffe2tnn slice layer convert error (#271)
  [CONVERTER][BUG] fix gemm bias's bug (#263)
  [ATLAS][BUG] the left pos of paste roi must be aligned by 16
  [ATLAS][CHG] refactor mat converter to seperate Resize and ResizeAndPaste
  [ATLAS][CHG] refactor altas test
  [WINDOWS] fix bugs with building on Windows system
  alloc first (#238)
  fix android demo memory leak (#258)
  [ATLAS][CHG] change AclInit param to avoid error info
  ...

+++ scripts/build_android.sh
@@ -11,6 +11,7 @@ OPENMP="ON"
 OPENCL="ON"
 BENMARK_MODE="OFF"
 DEBUG="OFF"
+INCREMENTAL_COMPILE="OFF"
 SHARING_MEM_WITH_OPENGL=0
 
 # check ANDROID_NDK whether set.
@@ -19,6 +20,19 @@ if [ ! -f "$ANDROID_NDK/build/cmake/android.toolchain.cmake" ]; then
    echo -e "Please download android ndk and set ANDROID_NDK environment variable."
    exit -1
 fi
+
+while [ "$1" != "" ]; do
+    case $1 in
+        -ic) 
+            shift
+            INCREMENTAL_COMPILE="ON"
+            ;;  
+        *)  
+            usage
+            exit 1
+    esac
+done
+
 
 TNN_BUILD_PATH=$PWD
 if [ -z $TNN_ROOT_PATH ]
@@ -37,14 +51,16 @@ source $TNN_VERSION_PATH/add_version_attr.sh
 
 echo ' '
 echo '******************** step 2: start build rpn arm32 ********************'
-#删除旧SDK文件
 cd $TNN_BUILD_PATH
 if [ -x "build32" ];then
-rm -r build32
+    if [ "${INCREMENTAL_COMPILE}" = "OFF" ];then
+        rm -r build32
+        mkdir build32
+    fi
+else
+    mkdir build32
 fi
 
-#新建build目录
-mkdir build32
 cd build32
 echo $ABIA32
 cmake ${TNN_ROOT_PATH} \
@@ -66,14 +82,16 @@ make -j32
 
 echo ' '
 echo '******************** step 3: start build rpn arm64 ********************'
-#删除旧SDK文件
 cd $TNN_BUILD_PATH
 if [ -x "build64" ];then
-rm -r build64
+    if [ "${INCREMENTAL_COMPILE}" = "OFF" ];then
+        rm -r build64
+        mkdir build64
+    fi  
+else
+    mkdir build64
 fi
 
-#新建build目录
-mkdir build64
 cd build64
 echo $ABIA64
 cmake ${TNN_ROOT_PATH} \
